backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
playwright==1.49.0
httpx==0.28.0
openai==1.58.0
anthropic==0.40.0
google-generativeai==0.8.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
